Add error reporting, rewrite GoodHypergraph file
GoodHypergraphCategory was taking a _very_ long time to compile
what's more, it triggered a bug in the compiler relating to how
indices are tracked in lambda declarations. This problem is supposed
to be resolved with #4851
(https://github.com/idris-lang/Idris-dev/pull/4851)

This bypasses the bug and improves compile times as well as add
error reporting on the TDef returned by the API in order to more
easily debug the source of failures when checking as petrinet
description.